+++ mitra_activitypub/src/adapters/mod.rs
@@ -1,2 +1,3 @@
+pub mod follow_requests;
 pub mod posts;
 pub mod users;