+++ mitra_api/src/mastodon_api/statuses/views.rs
@@ -276,6 +276,7 @@ async fn create_status(
     let post_data = PostCreateData {
         id: None,
         context: context,
+        title: status_data.title.clone(),
         content: content,
         content_source: content_source,
         language: status_data.language()?,
@@ -291,7 +292,7 @@ async fn create_status(
         object_id: None,
         created_at: Utc::now(),
     };
-    validate_post_create_data(&post_data)?;
+    validate_post_create_data(&post_data, Local)?;
     validate_post_mentions(&post_data.mentions, post_data.visibility)?;
     validate_local_post_links(&post_data.links, post_data.visibility)?;
     if let Some(ref in_reply_to) = maybe_in_reply_to {
@@ -509,6 +510,7 @@ async fn edit_status(
 
     // Update post
     let post_data = PostUpdateData {
+        title: status_data.title.clone(),
         content: content,
         content_source: content_source,
         language: status_data.language()?,
@@ -522,7 +524,7 @@ async fn edit_status(
         url: None,
         updated_at: Some(Utc::now()),
     };
-    validate_post_update_data(&post_data)?;
+    validate_post_update_data(&post_data, Local)?;
     validate_post_mentions(&post_data.mentions, post.visibility)?;
     validate_local_post_links(&post_data.links, post.visibility)?;
     if let Some(ref in_reply_to) = maybe_in_reply_to {