+++ mitra_cli/src/commands/post.rs
@@ -147,6 +147,7 @@ impl CreatePost {
                 object_id: None,
                 audience: Some(AP_PUBLIC.to_owned()),
             },
+            title: None,
             content: content,
             content_source: None,
             language: None,
@@ -162,7 +163,7 @@ impl CreatePost {
             object_id: None,
             created_at: self.created_at,
         };
-        validate_post_create_data(&post_data)?;
+        validate_post_create_data(&post_data, Local)?;
         check_post_limits(&config.limits.posts, &post_data.attachments, Local)?;
         let db_client = &mut **get_database_client(db_pool).await?;
         let post = create_post(db_client, author.id, post_data).await?;