+++ mitra_models/migrations/V0130__post__add_title.sql
@@ -1,0 +1,4 @@
+ALTER TABLE post ADD COLUMN title TEXT;
+
+DROP INDEX post_content_tsvector_simple_index;
+CREATE INDEX post_content_tsvector_simple_index ON post USING GIN (to_tsvector('simple', COALESCE(title, '') || ' ' || content));