+++ mitra_models/migrations/V0131__activitypub_collection_item__add_relationship_id.sql
@@ -1,0 +1,1 @@
+ALTER TABLE activitypub_collection_item ADD COLUMN relationship_id INTEGER REFERENCES relationship (id) ON DELETE CASCADE;