+++ CHANGELOG.md
@@ -6,9 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added `title` parameter to `/api/v1/statuses` form data.
+- Add `title` field to `Status` API entity if ActivityPub object has a `name`.
+- Added `default_fts_config` dynamic configuration parameter for changing text search settings.
+
 ### Changed
 
 - Changed MSRV to 1.82.0.
+- Treat scheduled outgoing activities where Ed25519 key is not included as invalid.
+- Validate 'ap' URI before making database query when serving portable objects.
+- Add actor to `followers` collection in database when creating follow relationship.
 
 ## [5.5.0] - 2026-06-10